+++ src/lsapy/__init__.py
@@ -1,7 +1,6 @@
 """A Python package for Land Suitability Analysis."""
 
 from lsapy.criteria import SuitabilityCriteria
-from lsapy.functions import *
 from lsapy.lsa import LandSuitabilityAnalysis
 
 __author__ = "Baptiste Hamon"